Restore integration test setup.
