--- lib/tests/integration/main.rs
@@ -1,5 +1,0 @@
-mod common;
-mod sdk_info;
-mod send_event;
-mod setup_complete_with_details;
-mod setup_complete_without_details;
